+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Alpha 1.3.5](https://github.com/Nevah5/wordlebot/releases/tag/a1.3.5) - 08.02.2022
 ### Added
 - `./src/modules/db.js` module
+- All database interactions into db module
 
 ## [Alpha 1.3.4](https://github.com/Nevah5/wordlebot/releases/tag/a1.3.4) - 07.02.2022
 ### Changed